Added timer module
